test-requirements.txt
odoo14-addon-sale-delivery-state @ git+https://github.com/oca/sale-workflow@refs/pull/1372/head#subdirectory=setup/sale_delivery_state
odoo14-addon-pos_backend_communication @ git+https://github.com/oca/pos@refs/pull/574/head#subdirectory=setup/pos_backend_communication
odoo14-addon-pos_backend_partner @ git+https://github.com/oca/pos@refs/pull/575/head#subdirectory=setup/pos_backend_partner
odoo14-addon-sale_invoice_policy @ git+https://github.com/oca/sale-workflow@refs/pull/1385/head#subdirectory=setup/sale_invoice_policy
openupgradelib @ git+https://github.com/OCA/openupgradelib.git
odoo14-addon-account_reconcile_restrict_partner_mismatch @ git+https://github.com/oca/account-reconcile@refs/pull/360/head#subdirectory=setup/account_reconcile_restrict_partner_mismatch
